// lessons/n38-48/src/utils/types/post.type.ts
export type PostType = {
  id: number;
  slug: string;
  url: string;
  title: string;
  content: string;
  image: string;
  thumbnail: string;
  status: string;
  category: string;
  publishedAt: string;
  updatedAt: string;
  userId: number;
};

export type PaginationType = {
  data: PostType[][];
  page: number;
  limit: number;
  count: number;
};

export type CreatePostRequestType = {
  title: string;
  content: string;
  category: string;
};